# Improve memory used by tag helper descriptors