--- a/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/AllowedChildTagDescriptor.cs
+++ b/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/AllowedChildTagDescriptor.cs
@@ -30,8 +30,9 @@ public abstract class AllowedChildTagDescriptor : IEquatable<AllowedChildTagDesc
                 TagHelperDiagnostics.AddDiagnostics(this, value);
                 _containsDiagnostics = true;
             }
-            else
+            else if (_containsDiagnostics)
             {
+                TagHelperDiagnostics.RemoveDiagnostics(this);
                 _containsDiagnostics = false;
             }
         }