--- a/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/BoundAttributeDescriptorComparer.cs
+++ b/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/BoundAttributeDescriptorComparer.cs
@@ -12,7 +12,7 @@ internal sealed class BoundAttributeDescriptorComparer : IEqualityComparer<Bound
     /// <summary>
     /// A default instance of the <see cref="BoundAttributeDescriptorComparer"/>.
     /// </summary>
-    public static readonly BoundAttributeDescriptorComparer Default = new BoundAttributeDescriptorComparer();
+    public static readonly BoundAttributeDescriptorComparer Default = new();
 
     private BoundAttributeDescriptorComparer()
     {
@@ -34,21 +34,35 @@ internal sealed class BoundAttributeDescriptorComparer : IEqualityComparer<Bound
             return false;
         }
 
-        return
-            string.Equals(descriptorX.Kind, descriptorY.Kind, StringComparison.Ordinal) &&
-            descriptorX.IsIndexerStringProperty == descriptorY.IsIndexerStringProperty &&
-            descriptorX.IsEnum == descriptorY.IsEnum &&
-            descriptorX.HasIndexer == descriptorY.HasIndexer &&
-            descriptorX.CaseSensitive == descriptorY.CaseSensitive &&
-            descriptorX.IsEditorRequired == descriptorY.IsEditorRequired &&
-            string.Equals(descriptorX.Name, descriptorY.Name, StringComparison.Ordinal) &&
-            string.Equals(descriptorX.IndexerNamePrefix, descriptorY.IndexerNamePrefix, StringComparison.Ordinal) &&
-            string.Equals(descriptorX.TypeName, descriptorY.TypeName, StringComparison.Ordinal) &&
-            string.Equals(descriptorX.IndexerTypeName, descriptorY.IndexerTypeName, StringComparison.Ordinal) &&
-            string.Equals(descriptorX.Documentation, descriptorY.Documentation, StringComparison.Ordinal) &&
-            string.Equals(descriptorX.DisplayName, descriptorY.DisplayName, StringComparison.Ordinal) &&
-            ComparerUtilities.Equals(descriptorX.BoundAttributeParameters, descriptorY.BoundAttributeParameters, BoundAttributeParameterDescriptorComparer.Default) &&
-            ComparerUtilities.Equals(descriptorX.Metadata, descriptorY.Metadata, StringComparer.Ordinal);
+        if (descriptorX.Kind != descriptorY.Kind ||
+            descriptorX.IsIndexerStringProperty != descriptorY.IsIndexerStringProperty ||
+            descriptorX.IsEnum != descriptorY.IsEnum ||
+            descriptorX.HasIndexer != descriptorY.HasIndexer ||
+            descriptorX.CaseSensitive != descriptorY.CaseSensitive ||
+            descriptorX.IsEditorRequired != descriptorY.IsEditorRequired ||
+            descriptorX.Name != descriptorY.Name ||
+            descriptorX.IndexerNamePrefix != descriptorY.IndexerNamePrefix ||
+            descriptorX.TypeName != descriptorY.TypeName ||
+            descriptorX.IndexerTypeName != descriptorY.IndexerTypeName ||
+            descriptorX.Documentation != descriptorY.Documentation ||
+            descriptorX.DisplayName != descriptorY.DisplayName)
+        {
+            return false;
+        }
+
+        if (!ComparerUtilities.Equals(descriptorX.BoundAttributeParameters, descriptorY.BoundAttributeParameters, BoundAttributeParameterDescriptorComparer.Default))
+        {
+            return false;
+        }
+
+        if (descriptorX.Metadata is MetadataCollection metadataX &&
+            descriptorY.Metadata is MetadataCollection metadataY &&
+            !metadataX.Equals(metadataY))
+        {
+            return false;
+        }
+
+        return ComparerUtilities.Equals(descriptorX.Metadata, descriptorY.Metadata, StringComparer.Ordinal);
     }
 
     public int GetHashCode(BoundAttributeDescriptor? descriptor)