--- a/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/BoundAttributeParameterDescriptor.cs
+++ b/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/BoundAttributeParameterDescriptor.cs
@@ -11,6 +11,33 @@ namespace Microsoft.AspNetCore.Razor.Language;
 
 public abstract class BoundAttributeParameterDescriptor : IEquatable<BoundAttributeParameterDescriptor>
 {
+    private enum Flags : byte
+    {
+        ContainsDiagnostics = 0x01,
+        IsEnum = 0x02,
+        IsStringProperty = 0x04,
+        IsBooleanProperty = 0x08,
+        CaseSensitive = 0x10
+    }
+
+    private Flags _flags;
+
+    private bool HasFlag(Flags flag) => (_flags & flag) != 0;
+    private void SetFlag(Flags flags) => _flags |= flags;
+    private void ClearFlag(Flags flags) => _flags &= ~flags;
+
+    private void SetOrClearFlag(Flags flag, bool value)
+    {
+        if (value)
+        {
+            SetFlag(flag);
+        }
+        else
+        {
+            ClearFlag(flag);
+        }
+    }
+
     protected BoundAttributeParameterDescriptor(string kind)
     {
         Kind = kind;
@@ -18,11 +45,23 @@ public abstract class BoundAttributeParameterDescriptor : IEquatable<BoundAttrib
 
     public string Kind { get; }
 
-    public bool IsEnum { get; protected set; }
+    public bool IsEnum
+    {
+        get => HasFlag(Flags.IsEnum);
+        protected set => SetOrClearFlag(Flags.IsEnum, value);
+    }
 
-    public bool IsStringProperty { get; protected set; }
+    public bool IsStringProperty
+    {
+        get => HasFlag(Flags.IsStringProperty);
+        protected set => SetOrClearFlag(Flags.IsStringProperty, value);
+    }
 
-    public bool IsBooleanProperty { get; protected set; }
+    public bool IsBooleanProperty
+    {
+        get => HasFlag(Flags.IsBooleanProperty);
+        protected set => SetOrClearFlag(Flags.IsBooleanProperty, value);
+    }
 
     public string Name { get; protected set; }
 
@@ -32,9 +71,31 @@ public abstract class BoundAttributeParameterDescriptor : IEquatable<BoundAttrib
 
     public string DisplayName { get; protected set; }
 
-    public bool CaseSensitive { get; protected set; }
+    public bool CaseSensitive
+    {
+        get => HasFlag(Flags.CaseSensitive);
+        protected set => SetOrClearFlag(Flags.CaseSensitive, value);
+    }
 
-    public IReadOnlyList<RazorDiagnostic> Diagnostics { get; protected set; }
+    public IReadOnlyList<RazorDiagnostic> Diagnostics
+    {
+        get => HasFlag(Flags.ContainsDiagnostics)
+            ? TagHelperDiagnostics.GetDiagnostics(this)
+            : Array.Empty<RazorDiagnostic>();
+
+        protected set
+        {
+            if (value?.Count > 0)
+            {
+                TagHelperDiagnostics.AddDiagnostics(this, value);
+                SetFlag(Flags.ContainsDiagnostics);
+            }
+            else
+            {
+                ClearFlag(Flags.ContainsDiagnostics);
+            }
+        }
+    }
 
     public IReadOnlyDictionary<string, string> Metadata { get; protected set; }
 
@@ -42,6 +103,11 @@ public abstract class BoundAttributeParameterDescriptor : IEquatable<BoundAttrib
     {
         get
         {
+            if (!HasFlag(Flags.ContainsDiagnostics))
+            {
+                return false;
+            }
+
             var errors = Diagnostics.Any(diagnostic => diagnostic.Severity == RazorDiagnosticSeverity.Error);
 
             return errors;