--- a/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/DefaultBoundAttributeDescriptor.cs
+++ b/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/DefaultBoundAttributeDescriptor.cs
@@ -1,8 +1,6 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Collections.Immutable;
-
 namespace Microsoft.AspNetCore.Razor.Language;
 
 internal sealed class DefaultBoundAttributeDescriptor : BoundAttributeDescriptor
@@ -19,7 +17,7 @@ internal sealed class DefaultBoundAttributeDescriptor : BoundAttributeDescriptor
         string? displayName,
         bool caseSensitive,
         BoundAttributeParameterDescriptor[] parameterDescriptors,
-        ImmutableDictionary<string, string> metadata,
+        MetadataCollection metadata,
         RazorDiagnostic[] diagnostics)
         : base(kind)
     {