… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/BoundAttributeParameterDescriptorComparer.cs
… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/DefaultBoundAttributeDescriptorBuilder.cs
… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/DefaultBoundAttributeParameterDescriptorBuilder.cs
… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/DefaultRequiredAttributeDescriptor.cs
… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/DefaultRequiredAttributeDescriptorBuilder.cs
… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/DefaultTagHelperDescriptor.cs
… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/DefaultTagHelperDescriptorBuilder.cs
… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/MetadataCollection.cs
… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/RequiredAttributeDescriptor.cs
… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/RequiredAttributeDescriptorComparer.cs
… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/Resources.resx
… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/TagHelperDescriptor.cs
… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/TagHelperDescriptorComparer.cs
… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/TagHelperDiagnostics.cs
… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/TagMatchingRuleDescriptor.cs
… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/TagMatchingRuleDescriptorComparer.cs
… src/Compiler/Microsoft.AspNetCore.Razor.Language/test/MetadataCollectionTests.cs
… src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Microbenchmarks/Remote/RemoteTagHelperDeltaProviderBenchmark.cs
… src/Shared/Microsoft.AspNetCore.Razor.Utilities.Shared/ListExtensions.cs
… src/Shared/Microsoft.AspNetCore.Razor.Utilities.Shared/PooledObjects/PooledList`1.cs
… src/Shared/Microsoft.AspNetCore.Razor.Utilities.Shared/ThreadSafeFlagOperations.cs