This is the first of several changes coming for tag helper descriptors to reduce their memory footprint. In this particular change. I've made tweaks to the memory used by the tag helper descriptors themselves. Future changes will promote sharing objects between tag helper descriptors, stop using a builder pattern for construction, and shrink the serialization payloads of tag helper descriptors.

Here is a summary of the changes:

1. Each object that makes up a tag helper descriptor has a field for storing diagnostics. This is a lot of wasted space since the majority of the objects never have diagnostics and those fields end up holding empty arrays. So, I've added a `ConditionalWeakTable` to store any diagnostics, similar to how green nodes work.
2. I've collapsed loads of boolean fields into single "Flags" fields.
3. I've cleaned up the tag helper descriptor comparers.
4. A new `MetadataCollection` class has been introduced to store metadata (i.e., name/value pairs) more efficiently. Previously, every object that allowed metadata had a `Dictionary<string, string>`, which is overkill since most objects only have 0, 1, or 2 name/value pairs in their metadata. The largest are rarely more than 8.

The `TagHelperSerializationBenchmark` shows good improvements with this change:

### Old (Main)

``` ini

BenchmarkDotNet=v0.13.2.2052-nightly, OS=Windows 11 (10.0.25314.1010)
Intel Core i7-8700 CPU 3.20GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET SDK=8.0.100-preview.1.23115.2
  [Host]     : .NET 7.0.4 (7.0.423.11508), X64 RyuJIT AVX2
  Job-NWIIOW : .NET 7.0.4 (7.0.423.11508), X64 RyuJIT AVX2
  Job-AENCQN : .NET Framework 4.8.1 (4.8.9139.0), X64 RyuJIT VectorSize=256

PowerPlanMode=00000000-0000-0000-0000-000000000000  

```
|                                    Method |        Job |            Toolchain |         Mean |      Error |     StdDev |       Median |          Min |          Max |      Gen0 |     Gen1 |     Gen2 |  Allocated |
|------------------------------------------ |----------- |--------------------- |-------------:|-----------:|-----------:|-------------:|-------------:|-------------:|----------:|---------:|---------:|-----------:|
| &#39;Razor TagHelper Roundtrip Serialization&#39; | Job-NWIIOW |             .NET 7.0 | 28,031.60 μs | 530.566 μs | 743.780 μs | 27,792.65 μs | 27,133.76 μs | 29,890.14 μs |  218.7500 | 156.2500 | 156.2500 | 15208378 B |
|           &#39;Razor TagHelper Serialization&#39; | Job-NWIIOW |             .NET 7.0 | 17,229.21 μs | 336.176 μs | 492.762 μs | 17,307.98 μs | 16,422.83 μs | 18,207.24 μs |  156.2500 | 156.2500 | 156.2500 |  9825705 B |
|         &#39;Razor TagHelper Deserialization&#39; | Job-NWIIOW |             .NET 7.0 | 10,332.30 μs | 133.565 μs | 111.533 μs | 10,378.79 μs | 10,110.54 μs | 10,479.46 μs |   62.5000 |        - |        - |  5366224 B |
|                  &#39;TagHelpers GetHashCode&#39; | Job-NWIIOW |             .NET 7.0 |     17.15 μs |   0.290 μs |   0.257 μs |     17.16 μs |     16.74 μs |     17.53 μs |         - |        - |        - |          - |
| &#39;Razor TagHelper Roundtrip Serialization&#39; | Job-AENCQN | .NET Framework 4.7.2 | 41,012.16 μs | 806.312 μs | 928.550 μs | 41,073.79 μs | 39,151.62 μs | 42,963.25 μs | 2000.0000 | 923.0769 | 846.1538 | 15568690 B |
|           &#39;Razor TagHelper Serialization&#39; | Job-AENCQN | .NET Framework 4.7.2 | 24,044.73 μs | 278.377 μs | 232.457 μs | 24,047.92 μs | 23,726.12 μs | 24,408.02 μs | 1218.7500 | 968.7500 | 968.7500 |  9838762 B |
|         &#39;Razor TagHelper Deserialization&#39; | Job-AENCQN | .NET Framework 4.7.2 | 13,302.67 μs | 193.039 μs | 171.124 μs | 13,318.43 μs | 12,906.24 μs | 13,503.71 μs |  906.2500 |  93.7500 |        - |  5718967 B |
|                  &#39;TagHelpers GetHashCode&#39; | Job-AENCQN | .NET Framework 4.7.2 |     17.86 μs |   0.194 μs |   0.172 μs |     17.88 μs |     17.59 μs |     18.21 μs |         - |        - |        - |          - |

### New

``` ini

BenchmarkDotNet=v0.13.2.2052-nightly, OS=Windows 11 (10.0.25324.1000)
Intel Core i7-8700 CPU 3.20GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET SDK=8.0.100-preview.1.23115.2
  [Host]     : .NET 7.0.4 (7.0.423.11508), X64 RyuJIT AVX2
  Job-VIYVKJ : .NET 7.0.4 (7.0.423.11508), X64 RyuJIT AVX2
  Job-RIMPBK : .NET Framework 4.8.1 (4.8.9139.0), X64 RyuJIT VectorSize=256

PowerPlanMode=00000000-0000-0000-0000-000000000000  

```
|                                    Method |        Job |            Toolchain |         Mean |      Error |     StdDev |       Median |          Min |          Max |      Gen0 |      Gen1 |     Gen2 |  Allocated |
|------------------------------------------ |----------- |--------------------- |-------------:|-----------:|-----------:|-------------:|-------------:|-------------:|----------:|----------:|---------:|-----------:|
| &#39;Razor TagHelper Roundtrip Serialization&#39; | Job-VIYVKJ |             .NET 7.0 | 22,612.16 μs | 429.280 μs | 401.548 μs | 22,595.55 μs | 21,885.28 μs | 23,237.06 μs |  218.7500 |  156.2500 | 156.2500 | 14499643 B |
|           &#39;Razor TagHelper Serialization&#39; | Job-VIYVKJ |             .NET 7.0 | 12,027.30 μs | 235.511 μs | 220.297 μs | 12,031.18 μs | 11,643.88 μs | 12,447.63 μs |  140.6250 |  140.6250 | 140.6250 |  9116647 B |
|         &#39;Razor TagHelper Deserialization&#39; | Job-VIYVKJ |             .NET 7.0 |  9,769.22 μs | 110.624 μs | 103.478 μs |  9,728.16 μs |  9,609.46 μs |  9,992.05 μs |   62.5000 |         - |        - |  5366224 B |
|                  &#39;TagHelpers GetHashCode&#39; | Job-VIYVKJ |             .NET 7.0 |     16.06 μs |   0.192 μs |   0.170 μs |     16.01 μs |     15.81 μs |     16.39 μs |         - |         - |        - |          - |
| &#39;Razor TagHelper Roundtrip Serialization&#39; | Job-RIMPBK | .NET Framework 4.7.2 | 31,654.39 μs | 116.745 μs |  97.488 μs | 31,678.72 μs | 31,396.56 μs | 31,776.04 μs | 1937.5000 | 1062.5000 | 937.5000 | 14857098 B |
|           &#39;Razor TagHelper Serialization&#39; | Job-RIMPBK | .NET Framework 4.7.2 | 13,850.41 μs | 113.766 μs | 106.416 μs | 13,845.35 μs | 13,727.15 μs | 14,062.66 μs | 1109.3750 | 1000.0000 | 984.3750 |  9127239 B |
|         &#39;Razor TagHelper Deserialization&#39; | Job-RIMPBK | .NET Framework 4.7.2 | 12,470.22 μs |  87.162 μs |  72.784 μs | 12,464.00 μs | 12,370.69 μs | 12,600.38 μs |  906.2500 |   93.7500 |        - |  5718964 B |
|                  &#39;TagHelpers GetHashCode&#39; | Job-RIMPBK | .NET Framework 4.7.2 |     17.33 μs |   0.336 μs |   0.437 μs |     17.19 μs |     16.50 μs |     18.32 μs |         - |         - |        - |          - |